backend/pyproject.toml
[project]
name = "finpilot-backend"
version = "0.1.0"
description = "FinPilot AI – Financial compliance and intelligence backend"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pdfplumber>=0.11.0",
    "python-dateutil>=2.9.0",
    "pymongo>=4.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "openai>=1.25.0",
    "langgraph>=1.1.8",
    "langchain-core>=1.3.0",
    "langchain-openai>=1.1.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
serve  = "main:start"
seed   = "finpilot.scripts.seed_demo_data:main"

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["finpilot"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "requests>=2.33.1",
]
